update: refactor dashboard into godutch
