+++ app/src/main/java/com/example/godutch/ui/godutch/GoDutchViewModel.java
@@ -1,14 +1,14 @@
-package com.example.godutch.ui.dashboard;
+package com.example.godutch.ui.godutch;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class GoDutchViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public GoDutchViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is dashboard fragment");
     }
