+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,13 +7,12 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
-
-    <item
         android:id="@+id/navigation_gallery"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/title_gallery" />
 
+    <item
+        android:id="@+id/navigation_godutch"
+        android:icon="@drawable/ic_baseline_attach_money_24"
+        android:title="@string/title_godutch" />
 </menu>
