+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -12,10 +12,10 @@
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.example.godutch.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        android:id="@+id/navigation_godutch"
+        android:name="com.example.godutch.ui.godutch.GoDutchFragment"
+        android:label="@string/title_godutch"
+        tools:layout="@layout/fragment_godutch" />
 
     <fragment
         android:id="@+id/navigation_gallery"
